kenteken-zoeker/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
rapidfuzz==3.9.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1